+++ Assets/TishPish/Scene/signalstatus.unity.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6837eb4d3b6ab894e82ab0d9a33ad011
+timeCreated: 1460815496
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
