+++ Assets/TishPish/Script/Showlayer.meta
@@ -1,6 +1,8 @@
 fileFormatVersion: 2
-guid: ea2a5216ffb8d4b319d09e8d7782120c
+guid: 934fe14843e5314408a98cf3b8d440f5
 folderAsset: yes
+timeCreated: 1460816061
+licenseType: Free
 DefaultImporter:
   userData: 
   assetBundleName:
