+++ Assets/TishPish/Script/SigStatScript.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ae4b3e21ea0cccb448e30f29ed41e81f
+timeCreated: 1460815499
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
